feat: add session-based local caching to reduce redundant Sugar-AI requests
