+++ solution/cache.py
